+++ src/fs/watchNodeFs.ts
@@ -2,6 +2,7 @@ import CheapWatch from 'cheap-watch';
 
 import type {PathStats, PathFilter} from './pathData.js';
 import {omitUndefined} from '../utils/object.js';
+import type {PartialExcept} from '../types.js';
 
 /*
