+++ src/index.ts
@@ -4,3 +4,5 @@ export type {Task} from './task/task.js';
 export {TaskError} from './task/task.js';
 
 export type {Gen} from './gen/gen.js';
+
+export * from './types.js';
