+++ src/project/rollup-plugin-gro-esbuild.ts
@@ -10,6 +10,7 @@ import {printKeyValue, printMs, printPath} from '../utils/print.js';
 import {toRootPath, isSourceId, TS_EXTENSION} from '../paths.js';
 import {omitUndefined} from '../utils/object.js';
 import {replaceExtension} from '../utils/path.js';
+import type {PartialExcept} from '../types.js';
 
 // TODO improve along with Svelte compile stats
 interface Stats {
