+++ src/project/rollup-plugin-output-css.ts
@@ -7,6 +7,7 @@ import {outputFile} from '../fs/nodeFs.js';
 import {SystemLogger, Logger} from '../utils/log.js';
 import type {GroCssBuild, GroCssBundle} from './types.js';
 import {omitUndefined} from '../utils/object.js';
+import type {PartialExcept} from '../types.js';
 
 export interface Options {
 	getCssBundles(): Map<string, GroCssBundle>;
