+++ src/project/rollup-plugin-plain-css.ts
@@ -7,6 +7,7 @@ import {green} from '../utils/terminal.js';
 import {SystemLogger} from '../utils/log.js';
 import type {GroCssBuild} from './types.js';
 import {omitUndefined} from '../utils/object.js';
+import type {PartialExcept} from '../types.js';
 
 export interface Options {
 	addCssBuild(build: GroCssBuild): boolean;
