+++ src/utils/object.ts
@@ -1,3 +1,5 @@
+import type {OmitStrict} from '../types.js';
+
 // Iterated keys in `for..in` are always returned as strings,
 // so to prevent usage errors the key type of `mapFn` is always a string.
 // Symbols are not enumerable as keys, so they're excluded.
